CpuId: Properly retrieve the Intel brand string
Follow the algorithm as in Figure 3-9 of [1] for the CPUID instruction. Text for the brand string was updated as per Table 3-14 of [1].

The check for the extended leaf wasn't checking for the most significant bit, thus potentially checking registers on processors that don't support the extended leaf.

References:
 [1] - Intel(R) 64 and IA-32 Architectures Software Developer's Manual Volume 2, Order Number 325383-072US, May 2020

Issue: DOTNET-275