CpuIdDll: Ensure that a dump occurs on the current core
Get the current processor the thread is running on, and set the affinity mask to that processor.

After reviewing AMD documentation, it appears that all "reserved" values are there to allow code to also run on AMD and to query. The information may still be useful, even though it's "reserved", so dump those reserved registers also.

Issue: DOTNET-280