CpuId: Add support for AMD
Detect the vendor string, and instantiate an object that decodes AMD CPUID information. The native DLL was updated to query 00000007h, 8000001Dh in a single dump.

Some refactoring for the features was moved from GenuineIntelCpu into GenericIntelCpuBase which is common for Intel and AMD.

Debug mode contains extra checks for features that ensure:
- A feature with the same name isn't added twice
- Reserved features are now added, so it's easier to see if there's a bit set which is unknown
- The same bit can't be used for more than one feature

APICId and APICMaxThreads were removed, as these aren't features, but required when calculating the geometry. On their own, they're not useful.

Intel features updated based on AMD documentation (function 0Dh).

Issue: DOTNET-274