CpuIdWin: Rework User Interface to allow more information
Later updates to CpuId will include Cache Topology and CPU Topology. There could be extra information added over time also. The current "tab" structure is not extensible, and will be difficult to maintain.

Implement a more object oriented approach, using a Tree structure on the left, which can contain more than one CPU node (so the program can query 'n' nodes, for 'n' different processors in the system). And then details for each node can be shown. And instead of having tabs, have subnodes in the tree, one for each block of information we'd like to show, such as register dumps, CPU caches, Processor Topology, etc. This would be much easier to extend allowing easier prototyping.

The window is updated to also be more compatible with Windows Vista and later, but still works well on Windows XP.

Issue: DOTNET-295