CpuId: Get all cores
Read the CPUID instruction for all cores, not just the current core. The DLL is extended to use OS specific functions to get the CPUID information for each core. This implementation only supports the current processor group in which the program is started.

Split x86CpuIdFactory functionality in three: a base (common code); one for reading registers (live data), the other for reading from XML.

The CpuIdXmlFactory is now responsible for writing data to XML. The X86CpuIdFactory handles the x86 specific code. This split is consistent the the loading functionality and layered similarly. Main code should call CpuIdXmlFactory.Save() to write the collection of ICpuId elements to disk.

Issue: DOTNET-296