CpuId: Refactor namespaces
The project is growing in size. Having everything in the RJCP.Diagnostics namespace clutters it, making it more difficult for users to find automatically generated documentation with SHFB. Instead, keep the most basic classes (factories) in the RJCP.Diagnostics namespace, and move everything else to the RJCP.Diagnostics.CpuId namespace.

Issue: DOTNET-303