CpuIdDll: Return compatibility to Windows XP
The CPUID.DLL library uses GetCurrentProcessorNumber, which is supported on Windows 2003 and later, but not in Windows XP. Thus, load the function dynamically. If it is supported, then use it, otherwise assume processor #0.

When iterating over the processors, use the System Information to know the processor mask, and iterate over the bitfield instead of being based on the maximum number of processors.

Issue: DOTNET-305