CpuId.DLL: AuthenticAMD Core Topology
A Node is an element specific to AMD, and is the Die on a package. Intel doesn't define the Die / Node, and the bits are likely different. At this time the Die is not used by Intel and so is effectively undefined. We can add the "Node" to Intel later if they have the same concept.

Issue: DOTNET-294