CpuId.DLL: Provide NuSpec package
The NuSpec package is a template, which is used by a private tool to generate automatically the NuGet package by inserting the version based on the revision control system. Otherwise, it can be easily modified to be used standalone.

It contains a targets file that will copy the DLL to the output directory. If creating a library that depends on CPUID, one needs to have the user also reference the assembly directly to get the native DLL. This appears to be a limitation of the NuGet packaging system. By not having the targets file, we can't copy the native library to the output at all. Having the Native DLL in the lib/net40 library will fail on adding the packages as references can't be directly added.

For reference, the SQLite library has the same problem, so it is not limited here, and so the current behaviour is considered acceptable, if not optimal.

Issue: DOTNET-269