CpuId.DLL: AMD for NPT Family 0Fh Model > 40h
The wrong powerlimit was calculated, and then misread the documentation that the value needs to be subtracted by 1, not added by 1.

Now the description and the brandstring match for these processors.

While reviewing the tables against #33610 Rev 3.48 against the implementation, noticed also the wrong power limit value for one entry.

Issue: DOTNET-306