CpuId.DLL: AMD for Family 10h, 11h, 12h, 14h
Compared the BIOS generated string Brand Description (80000002-4h) for families 10h, 11h, 12h, 14h with the BIOS documentation from AMD. Found discrepancies which showed an incorrect implementation in the .NET code in most cases (and some samples that had an incorrectly implemented BIOS).

Fixed interpretation of an "AMD Engineering Sample" that resulted in the wrong string. If the brand is zero, then it's an engineering sample.

In most places the partial mode was incorrectly calculated with minor errors (bit missing).

Issue: DOTNET-306