CpuIdWin: Enable HighDPI
The UserControl CpuIdTree needs to be set to AutoScaleMode.Inherit. If it isn't then .NET scales that control incorrectly when it is added to the main form (it is scaled correctly however during the construction).

Because the UserControl CpuIdTree is now AutoScaleMode.Inherit, the control now needs to recurse through the parents actually looking on how to scale properly for the other controls.

Issue: DOTNET-304