CpuId: Sign binaries with Authenticode
Use the SignFile task within MSBuild to sign executables. The <TargetsTriggeredByCompilation> is used so that signing only occurs when there's a change as suggested in [1]. A timestamp server is used when building the Release build only.

The signing of the binaries is only provided when the project is built on the command line with the user providing the signing certificate thumbprint as /p:CertificateThumbprint=XXXXXX (and the code signing certificate is available in Windows Certificate Store). The actual signing of the binaries is therefore not done with the VS-IDE, but instead, in an external script outside the scope of this project.

By signing the binaries external to this project means that the private key can change without changing this project.

The DLL is signed manually:

  "c:\Program Files (x86)\Windows Kits\10\Tools\bin\i386\signtool.exe" sign /fd sha1 /sha1 2fda16f7adf7153e17d4bf3d36adc514a736cdf4 /tr http://timestamp.digicert.com x64\cpuid.dll

Other projects aren't deployed and don't need signing.

Issue: DOTNET-310

[1] https://stackoverflow.com/questions/52496611/sign-file-msbuild-with-incremental-build