CpuID: Move to .NET SDK project format
Had to drop support for .NET 4.0 for CpuIdWin, as .NET SDK doesn't support this target framework for Windows. That means the GUI will only work on Windows 7 and later.

It will also compile under Linux using the same commands, although the CpuIdWin had to be removed from the project (to avoid the NETSDK1100 error). Test cases fail, but it can be copied onto Windows and it will work again.

Issue: DOTNET-343, DOTNET-350