CpuId: Avoid a race condition on loading CPUID.DLL
The `LoadLibrary` documentation states "if lpFileName does not include a path and there is more than one loaded module with the same base name and extension, the function returns a handle to the module that was loaded first", which this code relies on. However, if `FreeLibrary` is called, this can be unmapped, and a subsequent call to use a P/Invoke is called, the library will not be found.

This was tested by explicitly closing the handle on Windows 10.19043 after the `LoadLibrary`, where the test cases then failed that the library name could not be found.

Thus, the `SafeLibraryHandle` must be made static that it is not finalized, avoiding a potential race condition.