CpuId: Rework P/Invoke to use best practices
Native Win32 APIs were updated so:

* arguments that are `out` are replaced from `ref`.
* attribute `ExactSpelling` is used
* a bug with `LoadLibrary` was fixed to use the correct character set
* The `OSArchitecture` is made a `short`, so that it is two bytes and can be used directly in `PROCESSOR_INFO_UNION` without having to type cast.

When running in Debug Mode on x64 with .NET 4.0, received an `InvalidProgramException` at the P/Invoke of getting the CPUID dump for our native library. This indicates some marshalling error. The API was changed to use unsafe pointers instead of arrays, which works properly also in debug mode. Note, that .NET Core didn't show the error, so it might have been a problem in .NET Framework only.

Issue: DOTNET-716