CPUID: Add information to features
Simplified the testing of bits for features in code, and corrected some errors.

The `CpuFeatures` now returns a `CpuFeature` class which provides information if the feature is available (instead of a `bool`), along with the bit and the feature group. This can be used to provide additional information to the user to identify the precise function from the chip manufacturer's datasheet.

Issue: DOTNET-883