CPUID AMD: Fix feature bits to use names from AMD documentation
The names are taken from the CPUID descriptions of [1]. Aliases to Intel bit names are added where functionality overlaps.

Aliases for an AMD CPU are:
* "SSE4_1" => AMD "SSE41"
* "SSE4_2" => AMD "SSE42"
* "AESNI" => AMD "AES"
* "CX8" => AMD "CMPXCHG8B"
* "SEP" => AMD "SysEnterSysExit"
* "PSE-36" => AMD "PSE36"
* "VPCLMULQDQ" => AMD "VPCMULQDQ"
* "BUS_LOCK_DETECT" => AMD "BUSLOCKTRAP"
* "AHF64" => AMD "LahfSahf"
* "LZCNT" => AMD "ABM"
* "PREFETCHW" => AMD "3DNowPrefetch"
* "SYSCALL" => AMD "SysCallSysRet"
* "XD" => AMD "NX"
* "1GB_PAGE" => AMD "Page1GB"

Intel features renamed:
* Intel's ABM bit was renamed to LZCNT to match the Intel documentation. On AMD, it remains ABM.
* SSE4.1 => SSE4_1
* SSE4.2 => SSE4_2

By introducing the feature of aliases, test cases were updated, such that they must check for the CPU name, not the aliased name. The check `AssertOnMissingDescription` was updated add add aliases as known features. CpuIdWin was updated to not print aliases (they'll now show up in the enumerated list of `CpuFeatures` multiple times).

[1] "AMD64 Architecture Programmer’s Manual Volume 3: General-Purpose and System Instructions", Publication 24594, Rev 3.35, June 2023.

Issue: DOTNET-887