CPUID DLL: Reimplement in C++17
Improved the maintainability of the CPUID DLL library by rewriting in C++17. Uses modern constructs of OOP to make it easier to extend for other CPU types.

While there are no test cases for this library, it should be easier to implement and test the individual classes. See `DESIGN.md` for a breakdown of how this (simple) library is implemented and some of the decisions made.

It should now be easier to extend for functionality in newer versions of Windows (e.g. 64-bit with more than 64 processors) through the `Kernel32` class. This version doesn't support more than 64 processors on a 64-bit machine however.

Visual Studio 2022 is used for this development and the binary is still targetted for Windows XP. Much of the original code was taken from another repository of mine `devc-cpuid` which is still a work in progress.
* clang-tidy is enabled for MSVC 2022, which allowed for testing correctness of the library at compile time, even if the static analysers are not implemented for Windows XP based code.
* clang-format is used to create a style similar to Google's coding style. Some minor differences are made, to allow for wider columns. For formatting class information, CodeMaid is used with 120 column break (and the clang-format is configured for this too). The include order is tweaked so that `stdafx.h` is always on the top as it must be (it includes `<Windows.h>`).

Testing the Library:
* The `cpdll.bat` was simplified to make it easier to copy the file for test cases also. On Windows, I use links which makes it easier to compile and debug even if the test cases are in .NET. See `DEBUG.md` on how I used .NET Framework test cases with NUnit.
* The idea is that the globalstate contains the factory to generate the CPUID information. This can be replaced with a test case implementation that could provide other sources of CPUID (e.g. such as a file).

Issue: DOTNET-888