CPUID DLL: Use alternatives to `GetCurrentProcessorNumber()`
Windows XP doesn't have the Kernel32 call `GetCurrentProcessorNumber()`. Use two possible alternatives to get the CPUID for the current thread (that requires pinning):

* Use `ntdll.dll` call `NtGetCurrentProcessorNumber()` if it exists; else
* Use the CPUID APIC-ID as a replacement for `GetCurrentProcessorNumber()`

The CPUID solution iterates through all available hardware threads and gets the APIC-ID. Then when we're asked what thread we're on, we get the APICID using the CPUID instruction and map that back to the original thread. Then we know which bit to use when setting the affinity to stay on that thread.

Note, we only use the 8-bit variant of the APIC-ID. Intel documentation says we should use CPUID.1Fh or CPUID.0Bh to get the 32-bit APIC-ID, but this requires multiple calls, which is not acceptable (there might be a context switch in-between).

Issue: DOTNET-309