CpuID: Introduce specifying the core
For `CpuIdLibFactory`, the methods `ICpuIdFactory.Create()` and `ICpuRegisters.GetCpuId()` operate on whatever core the system is running on. This results in undefined behaviour. It works because the `ICpuIdFactory.CreateAll()` does pin threads and collects for all cores, thus meaning the `Create()` and `GetCpuId()` methods are generally not called at all.

For `CpuIdXmlFactory`, it would always return the first core, so was well defined.

The `ICpuIdFactory.Create()` should now create from the first core so the results are defined (instead of the current core, which is non-portable on .NET, so we implement that in this commit). Add a new method `ICpuIdFactory.Create(int core)` to specify a core to get information for. Then the `ICpuRegisters.GetCpuId()` should run on the core that was given to the factory method, thus the results are well defined.

The `CpuIdXmlFactory.CreateAll(string)` is now made static, as it isn't needed to be a member class, it isn't required as part of the factory interface `ICpuIdFactory`.

Issue: DOTNET-1061