CpuId: Separate the factory and the ICpuRegisters
The capture of CPUID information for a single core is moved from `CpuIdLibRegisters` to `CpuIdLibFactory`, so that common functionality is in a single class, instead of being spread over two classes. The implementation of `CpuIdLibRegisters` is reduced to only obtaining a CPU register only if the initial enumeration didn't capture the value.

Same that `CpuIdNetRegisters` is reduced to only obtaining a CPU register. It allows the factory to pin to a core and query without pinning as a simple performance optimisation.

The `CpuIdX86XmlRegisters` is no longer needed as it can't get CPUID information. Reading the XML file is now done completely in the factory `CpuIdX86CmlNodeFactory`.

The `ICpuRegisters` is no longer enumerable, it is only for capturing CPUID information for a specific core.

The user is exposed now only `CpuRegisters` which remains enumerable, as this implementation caches the captured CPUID information, which is obtained by the factory class.

Issue: DOTNET-1038