CpuIdWin: Load images explicitly
When compiling, the warning shows:

> Resource "imgList.ImageStream" will be deserialized via BinaryFormatter at runtime. BinaryFormatter is deprecated due to known security risks and is removed from .NET 9+. If you wish to continue using it, set property "GenerateResourceWarnOnBinaryFormatterUse" to false.

The issue [1] makes it clear that the BinaryFormatter is deprecated. So the warnings and workarounds are to be removed from the `.csproj` file.

Despite [2] that describes backwards compatibility, this only appears to be the case for a main form. But our `imgList` is on a custom control, which is loaded by the main form. Opening the custom control in the WinForms editor of VS2022/2026 works fine, but not when opening the form that contains the custom control.

Issue: DOTNET-1083

[1]: https://github.com/dotnet/winforms/issues/9701
[2]: https://github.com/dotnet/msbuild/issues/11185